Update operator version
We depend on the storageoscluster CRD in the operator repo.
Using an outdated version of such causes us to parse only the fields that exist on that version and ignore any new fields that would have been added to the CRD in the latest changes. This was happening with the metrics attribute causing it to be ignored and the next time we applied the change the field would reset disabling the metrics feature.